ajout de l'affichage par scale et rotation changement du type de scale (int à float) changement du type de rotation (int ) float) ajout de la configuration width et height sur une map

+++ src/assets-manager/config-manager/image-config-manager.c
@@ -299,7 +299,7 @@ bool loadRotation(ImageConfig* inConfig,yaml_parser_t* parser,char* parentDirPat
                 if(!nextIsValue)
                     break;
 
-                inConfig->rotation = atoi((char*)token.data.scalar.value);
+                inConfig->rotation = atof((char*)token.data.scalar.value);
                 yaml_token_delete(&token);
                 return true;
